테이블 프레임 분리 클래스
