+++ src/table_demo/GUIMain.java
@@ -47,183 +47,9 @@ public class GUIMain {
         // 거기서 GUI를 생성하고 보여준다.
         javax.swing.SwingUtilities.invokeLater(new Runnable() {
             public void run() {
-                //GUIMain.getInstance().createAndShowGUI();
-                GUIMain.getInstance().createAndShowReception();
-                GUIMain.getInstance().createAndShowPatient();
-                GUIMain.getInstance().createAndShowReservation();
+
             }
         });
-    }
-    /**
-     * GUI를 생성하여 보여준다. 스레드 안전을 위하여
-     * 이 메소드는 이벤트 처리 스레드에서 불려져야 한다.
-     */
-    static JFrame mainFrame = new JFrame("병원 환자 관리 프로그램");
-    JFrame receptionFrame = new JFrame("진료");
-    JFrame patientFrame = new JFrame("환자 정보");
-    JFrame reservationFrame = new JFrame("예방접종 예약");
-
-    /*private void createAndShowGUI() {
-        //mainFrame.setLocationRelativeTo(null);
-        mainFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        // 탭을 생성하고 두개 패널을 추가한다.
-        JTabbedPane jtab = new JTabbedPane();
-        jtab.setBackground(Color.WHITE);
-        //setupReceptionPane();
-        //setupPatientPane();
-        //setupVaccinationPane();
-        setupResevationPane();
-        // 아이템 리스트 탭과 주문 탭 두 개의 패널을 가지는 탭 패널
-        //jtab.add("접수", receptionPane);
-        //jtab.add("환자", patientPane);
-        //jtab.add("예방접종현황", vaccinationPane);
-        //jtab.add("예방접종예약", reservationPane);
-        mainFrame.getContentPane().add(jtab);
-        mainFrame.getContentPane().setBackground(Color.WHITE);
-        //Display the window.
-        mainFrame.pack();
-        mainFrame.setVisible(true);
-    }*/
-
-    //진료 프레임
-   private void createAndShowReception() {
-    	setupReceptionPane();
-
-		receptionFrame.setBounds(100, 100, 720, 600);
-		receptionFrame.setBackground(Color.WHITE);
-    	receptionFrame.add(receptionPane);
-    	receptionFrame.pack();
-    	receptionFrame.setVisible(true);
-    }
-    // 접수을 보여주는 패널 부분 - 탑과 JTable 포함
-    public JPanel receptionPane;
-    TableSelectionDemo receptionTable = new TableSelectionDemo();
-    SearchTopPanel receptionTop = new SearchTopPanel();
-    ReceptionDownPanel receptionDown = new ReceptionDownPanel();
-    private void setupReceptionPane() {
-        receptionPane = new JPanel(new BorderLayout());
-        receptionPane.setPreferredSize(new Dimension(720,600));
-        receptionPane.setBackground(Color.WHITE);
-
-        receptionTop.setBackground(new Color(147, 251, 206));
-        receptionDown.setBackground(Color.WHITE);
-        //Create and set up the content pane.
-        receptionTable.tableTitle = "reception";
-        receptionTable.addComponentsToPane(ReceptionMgr.getInstance());  // 싱글톤
-        receptionTop.setupTopPane(receptionTable);
-        receptionPane.add(receptionTop, BorderLayout.NORTH);
-        receptionPane.add(receptionTable, BorderLayout.CENTER);
-
-        receptionDown.setupDownPane(receptionTable);
-        receptionPane.add(receptionDown, BorderLayout.SOUTH);
-    }
-
-    //환자 프레임
-    private void createAndShowPatient() {
-    	JTabbedPane patientJtab = new JTabbedPane();
-    	patientJtab.setBackground(Color.WHITE);
-
-    	setupPatientPane();
-    	setupVaccinationPane();
-
-    	patientJtab.add("환자", patientPane);
-    	patientJtab.add("예방접종현황", vaccinationPane);
-
-		patientFrame.setBounds(100, 100, 720, 600);
-		patientFrame.setBackground(Color.WHITE);
-		patientFrame.add(patientJtab);
-		patientFrame.pack();
-		patientFrame.setVisible(true);
-    }
-    // 환자을 보여주는 패널 부분 - 위에는 검색과 JTable, 아래는 환자 등록
-    private JPanel patientPane;
-    TableSelectionDemo patientTable = new TableSelectionDemo();
-    ReceptionOfPatientTableDemo rpListTable = new ReceptionOfPatientTableDemo();
-    SearchTopPanel patientTop = new SearchTopPanel();
-    PatientDownPanel patientDown = new PatientDownPanel();
-    private void setupPatientPane() {
-        patientPane = new JPanel(new BorderLayout());
-        patientPane.setPreferredSize(new Dimension(720,600));
-
-        patientTop.setBackground(new Color(147, 251, 206));
-        patientDown.setBackground(Color.WHITE);
-
-        patientTable.tableTitle = "ReceptionOfPatient";
-        patientTable.addComponentsToPane(PatientMgr.getInstance());
-        patientTable.setBounds(0,0,720,299);
-
-        //맨 위에는 검색 창
-        patientTop.setupTopPane(patientTable);
-        patientPane.add(patientTop, BorderLayout.NORTH);
-        //가운데 환자 리스트와 진료 기록 테이블 추가
-        JPanel center = new JPanel();
-        center.setBackground(Color.WHITE);
-        patientPane.add(center, BorderLayout.CENTER);
-        center.setLayout(null);
-        //맨 아래는 환자 등록 창
-        patientDown.setupDownPane(patientTable);
-        patientPane.add(patientDown, BorderLayout.SOUTH);
-
-        //가운데 환자 리스트와 진료 기록 테이블 추가
-        center.add(patientTable);
-        rpListTable.tableTitle = "ReceptionOfPatientList";
-        rpListTable.addComponentsToPane(ReceptionOfPatientMgr.getInstance());
-        rpListTable.setBounds(0, 300, 720, 205);
-        center.add(rpListTable);
-    }
-    //예방접종 현황 탭
-    private JPanel vaccinationPane;
-    TableSelectionDemo v_patientTable = new TableSelectionDemo();
-    ReservationOfPatientTableDemo v_rListTable = new ReservationOfPatientTableDemo();
-    VaccinationTopPanel v_patientTop = new VaccinationTopPanel();
-    private void setupVaccinationPane() {
-        //검색
-        vaccinationPane = new JPanel(new BorderLayout());
-        v_patientTop.setBackground(new Color(147, 251, 206));
-
-        v_patientTop.setupTopPane(v_patientTable);
-        vaccinationPane.add(v_patientTop, BorderLayout.NORTH);
-
-        v_patientTable.tableTitle = "ReservationOfPatient";
-        v_patientTable.addComponentsToPane(PatientMgr.getInstance());
-        vaccinationPane.add(v_patientTable, BorderLayout.CENTER);
-
-        v_rListTable.tableTitle = "ReservationOfPatientList";
-        v_rListTable.addComponentsToPane(ReservationOfPatientMgr.getInstance());
-
-        vaccinationPane.add(v_rListTable, BorderLayout.SOUTH);
-    }
-
-    //예방접종예약 프레임
-    private void createAndShowReservation() {
-    	setupResevationPane();
-
-    	reservationFrame.setBounds(100, 100, 720, 600);
-    	reservationFrame.setBackground(Color.WHITE);
-    	reservationFrame.add(reservationPane);
-    	reservationFrame.pack();
-    	reservationFrame.setVisible(true);
-    }
-    //예방접종 예약 탭
-    private JPanel reservationPane;
-    TableSelectionDemo reservationTable = new TableSelectionDemo();
-    SearchTopPanel reservationTop = new SearchTopPanel();  // 검색과 상세보기 버튼을 가진 패널
-    ReservationDownPanel reservationDown = new ReservationDownPanel();
-    private void setupResevationPane() {
-        reservationPane = new JPanel(new BorderLayout());
-        reservationPane.setPreferredSize(new Dimension(720,600));
-
-        reservationTop.setBackground(new Color(147, 251, 206));
-        reservationDown.setBackground(Color.WHITE);
-        //Create and set up the content pane.
-        reservationTable.tableTitle = "reservation";
-        reservationTable.addComponentsToPane(ReservationMgr.getInstance());  // 싱글톤
-        reservationTop.setupTopPane(reservationTable);
-        reservationPane.add(reservationTop, BorderLayout.NORTH);
-        reservationPane.add(reservationTable, BorderLayout.CENTER);
-
-        reservationDown.setupDownPane(reservationTable);
-        reservationPane.add(reservationDown, BorderLayout.SOUTH);
     }
 
 }
