+++ src/table_demo/TableSelectionDemo.java
@@ -83,23 +83,21 @@ public class TableSelectionDemo extends JPanel implements ListSelectionListener 
         if (!lsm.isSelectionEmpty()) {
             selectedIndex = lsm.getMinSelectionIndex();
             String name = (String)tableModel.getValueAt(selectedIndex, 1);
-            // 아이템 테이블의 클릭은 텍스트 필드에 값을 보여주고
-            // 주문 테이블의 클릭은 장바구니의 값을 바꾼다
+
             if (tableTitle.equals("reception")) {
-                GUIMain.getInstance().receptionTop.kwdTextField.setText(name);
+                ReceptionTab.getInstance().receptionTop.kwdTextField.setText(name);
             }
             else if(tableTitle.equals("ReceptionOfPatient")) {
-                GUIMain.getInstance().patientTop.kwdTextField.setText(name);
-                GUIMain.getInstance().rpListTable.loadData(""+selectedIndex);
+                PatientTab.getInstance().patientTop.kwdTextField.setText(name);
+                PatientTab.getInstance().rpListTable.loadData(""+selectedIndex);
             }
             else if(tableTitle.equals("ReservationOfPatient")) {
-                GUIMain.getInstance().v_patientTop.kwdTextField.setText(name);
-                GUIMain.getInstance().v_rListTable.loadData(""+selectedIndex);
+                PatientTab.getInstance().v_patientTop.kwdTextField.setText(name);
+                PatientTab.getInstance().v_rListTable.loadData(""+selectedIndex);
             }
             else if(tableTitle.equals("reservation")) {
-                GUIMain.getInstance().reservationTop.kwdTextField.setText(name);
+                ReservationTab.getInstance().reservationTop.kwdTextField.setText(name);
             }
-
         }
     }
 }
